[pytx] Move storage interface to the right place
